+++ rs-engine/src/active_npc.rs
@@ -198,7 +198,7 @@ impl ActiveNpc {
     /// * When `reset` is true, adjusts `npc.levels` and `npc.base_levels` to
     ///   match the new type's stats.
     /// * Schedules a revert at `clock + duration` (stored in `npc.revert_at`).
-    pub fn change_type(&mut self, new_type: u16, duration: u64, reset: bool, clock: u64) {
+    pub fn change_type(&mut self, new_type: u16, duration: u64, reset: bool, clock: u32) {
         if duration < 1 {
             return;
         }
@@ -226,7 +226,7 @@ impl ActiveNpc {
         if new_type == self.npc.base_type && self.npc.lifecycle == EntityLifeTime::Respawn {
             self.npc.revert_at = None;
         } else {
-            self.npc.revert_at = Some(clock + duration);
+            self.npc.revert_at = Some(clock + duration as u32);
             self.npc.revert_reset = reset;
         }
     }
